HHH-19846 Drop JUnit 4 usage
# Conflicts:
#	hibernate-core/src/test/java/org/hibernate/orm/test/entitymode/map/basic/DynamicClassTest.java
#	hibernate-core/src/test/java/org/hibernate/orm/test/fetching/GraphParsingTest.java

+++ hibernate-core/src/test/java/org/hibernate/orm/test/compositeelement/Child.java
@@ -70,6 +70,6 @@ public class Child {
 		return bioLength;
 	}
 	public void setBioLength(Integer bioLength) {
-		this.bioLength = bioLength==null ? 0 : bioLength.intValue();
+		this.bioLength = bioLength==null ? 0 : bioLength;
 	}
 }
